refactor sub, move stuff out to common
