+++ src/Parser/Rogue/Assasinaition/CHANGELOG.js
@@ -1,0 +1,3 @@
+export default `
+
+`;
