+++ src/Parser/Rogue/Assasinaition/CONFIG.js
@@ -1,0 +1,14 @@
+import SPECS from 'common/SPECS';
+import SPEC_ANALYSIS_COMPLETENESS from 'common/SPEC_ANALYSIS_COMPLETENESS';
+
+import CombatLogParser from './CombatLogParser';
+import CHANGELOG from './CHANGELOG';
+
+export default {
+  spec: SPECS.ASSASSINATION_ROGUE,
+  maintainers: [],
+  completeness: SPEC_ANALYSIS_COMPLETENESS.NEEDS_MORE_WORK, // good = it matches most common manual reviews in class discords, great = it support all important class features
+  changelog: CHANGELOG,
+  parser: CombatLogParser,
+  path: __dirname, // used for generating a GitHub link directly to your spec
+};
