+++ src/Parser/Rogue/Common/CastTracker/FilteredDamageTracker.js
