+++ src/Parser/Rogue/Common/Legendaries/DreadlordsDeceit.js
