+++ src/Parser/Rogue/Common/Legendaries/InsigniaOfRavenholdt.js
