+++ src/Parser/Rogue/Common/Legendaries/MantleDamageTracker.js
