+++ src/Parser/Rogue/Common/Legendaries/MantleOfTheMasterAssassin.js
