+++ src/Parser/Rogue/Common/Legendaries/SoulOfTheShadowblade.js
