+++ src/Parser/Rogue/Common/Resources/EnergyDetails.js
@@ -4,11 +4,7 @@ import Analyzer from 'Parser/Core/Analyzer';
 import Tab from 'Main/Tab';
 import StatisticBox, { STATISTIC_ORDER } from 'Main/StatisticBox';
 import Icon from 'common/Icon';
-import SPELLS from 'common/SPELLS';
-import SpellLink from 'common/SpellLink';
-import Wrapper from 'common/Wrapper';
 import ResourceBreakdown from 'Parser/Core/Modules/ResourceTracker/ResourceBreakdown';
-import resourceSuggest from 'Parser/Core/Modules/ResourceTracker/ResourceSuggest';
 
 import EnergyTracker from './EnergyTracker';
 
@@ -16,25 +12,6 @@ class EnergyDetails extends Analyzer {
   static dependencies = {
     energyTracker: EnergyTracker,
   };
-  
-  suggestions(when) {
-    resourceSuggest(when,  this.energyTracker, {
-      spell: SPELLS.SYMBOLS_OF_DEATH,
-      minor: 0.10,
-      avg: 0.20, 
-      major: 0.50,
-      extraSuggestion: <Wrapper>Try to spend energy before using <SpellLink id={SPELLS.SYMBOLS_OF_DEATH.id} />, but do not delay it to avoid waste! </Wrapper>,
-    });
-      
-    resourceSuggest(when,  this.energyTracker, {
-      spell: SPELLS.RELENTLESS_STRIKES,
-      minor: 0.15,
-      avg: 0.25, 
-      major: 0.40,
-      extraSuggestion: <Wrapper> You are wasting more energy then normal. You may be pooling too much energy or not casting enough spenders. </Wrapper>,
-    });
-  }
-
 
   statistic() {
     const energyWasted = this.energyTracker.wasted;
