+++ src/Parser/Rogue/Common/Resources/EnergyTracker.js
