+++ src/Parser/Rogue/Subtlety/CombatLogParser.js
@@ -6,10 +6,12 @@ import AlwaysBeCasting from './Modules/Features/AlwaysBeCasting';
 
 import ShadowBladesUptime from "./Modules/Features/ShadowBladesUptime";
 import SymbolsOfDeathUptime from "./Modules/Features/SymbolsOfDeathUptime";
-import ComboPointDetails from './Modules/RogueCore/ComboPointDetails';
-import ComboPointTracker from './Modules/RogueCore/ComboPointTracker';
-import EnergyDetails from './Modules/RogueCore/EnergyDetails';
-import EnergyTracker from './Modules/RogueCore/EnergyTracker';
+import ComboPointDetails from '../Common/Resources/ComboPointDetails';
+import ComboPointTracker from '../Common/Resources/ComboPointTracker';
+import ComboPoints from './Modules/RogueCore/ComboPoints';
+import EnergyDetails from '../Common/Resources/EnergyDetails';
+import EnergyTracker from '../Common/Resources/EnergyTracker';
+import Energy from './Modules/RogueCore/Energy';
 import SymbolsDamageTracker from './Modules/RogueCore/SymbolsDamageTracker';
 import DanceDamageTracker from './Modules/RogueCore/DanceDamageTracker';
 import DarkShadowNightblade from './Modules/Talents/DarkShadow/DarkShadowNightblade';
@@ -18,17 +20,17 @@ import DarkShadowEvis from "./Modules/Talents/DarkShadow/DarkShadowEvis";
 import DeathFromAbove from "./Modules/Talents/DFA/DeathFromAbove";
 import NightbladeDuringSymbols from './Modules/BaseRotation/NightbladeDuringSymbols';
 import CastsInShadowDance from './Modules/BaseRotation/CastsInShadowDance';
-import MantleDamageTracker from './Modules/Legendaries/MantleDamageTracker';
+import MantleDamageTracker from '../Common/Legendaries/MantleDamageTracker';
 import DeathFromAboveMantle from './Modules/Talents/DFA/DeathFromAboveMantle';
 import T21_2P from './Modules/Items/T21_2P';
 import DanceCooldownReduction from './Modules/RogueCore/DanceCooldownReduction';
 import DenialOfHalfGiants from './Modules/Legendaries/DenialOfHalfGiants';
 import FirstOfTheDead from './Modules/Legendaries/FirstOfTheDead';
-import MantleOfTheMasterAssassin from './Modules/Legendaries/MantleOfTheMasterAssassin';
+import MantleOfTheMasterAssassin from '../Common/Legendaries/MantleOfTheMasterAssassin';
 import ShadowSatyrsWalk from './Modules/Legendaries/ShadowSatyrsWalk';
-import SoulOfTheShadowblade from './Modules/Legendaries/SoulOfTheShadowblade';
-import InsigniaOfRavenholdt from './Modules/Legendaries/InsigniaOfRavenholdt';
-import DreadlordsDeceit from './Modules/Legendaries/DreadlordsDeceit';
+import SoulOfTheShadowblade from '../Common/Legendaries/SoulOfTheShadowblade';
+import InsigniaOfRavenholdt from '../Common/Legendaries/InsigniaOfRavenholdt';
+import DreadlordsDeceit from '../Common/Legendaries/DreadlordsDeceit';
 
 import DarkShadowSpecterOfBetrayal from './Modules/Talents/DarkShadow/DarkShadowSpecterOfBetrayal';
 
@@ -44,8 +46,10 @@ class CombatLogParser extends CoreCombatLogParser {
     //Resource
     comboPointTracker: ComboPointTracker,
     comboPointDetails: ComboPointDetails,
+    comboPoints: ComboPoints,
     energyTracker: EnergyTracker,
     energyDetails: EnergyDetails,
+    energy: Energy,
 
     //Trackers
     symbolsDamageTracker: SymbolsDamageTracker,
