+++ src/Parser/Rogue/Subtlety/Modules/RogueCore/ComboPoints.js
@@ -1,19 +1,15 @@
 import React from 'react';
 
 import Analyzer from 'Parser/Core/Analyzer';
-import Tab from 'Main/Tab';
-import StatisticBox, { STATISTIC_ORDER } from 'Main/StatisticBox';
 import SPELLS from 'common/SPELLS';
-import Icon from 'common/Icon';
 import SpellLink from 'common/SpellLink';
 import Wrapper from 'common/Wrapper';
-import ResourceBreakdown from 'Parser/Core/Modules/ResourceTracker/ResourceBreakdown';
 import resourceSuggest from 'Parser/Core/Modules/ResourceTracker/ResourceSuggest';
 
-import ComboPointTracker from './ComboPointTracker';
+import ComboPointTracker from '../../../Common/Resources/ComboPointTracker';
 
 
-class ComboPointDetails extends Analyzer {
+class ComboPoints extends Analyzer {
   static dependencies = {
     comboPointTracker: ComboPointTracker,
   };
@@ -68,37 +64,6 @@ class ComboPointDetails extends Analyzer {
       extraSuggestion: <span> Cast <SpellLink id={SPELLS.GOREMAWS_BITE.id}/> when you are on or below 3 combo points </span>,
     });
     }
-
-  statistic() {
-    const pointsWasted = this.comboPointTracker.wasted;
-    const pointsWastedPerMinute = (pointsWasted / this.owner.fightDuration) * 1000 * 60;
-    return (
-      <StatisticBox
-        icon={<Icon icon="ability_rogue_masterofsubtlety" alt="Waisted Combo Points" />}
-        value={`${pointsWastedPerMinute.toFixed(2)}`}
-        label="Wasted Combo Points per minute"
-        tooltip={`You waisted a total of ${pointsWasted} combo points. Some waste is expected due to the random nature of some generation abilities.`}
-      />
-    );
-  }
-
-  tab() {
-    return {
-      title: 'Combo Point usage',
-      url: 'combo-points',
-      render: () => (
-        <Tab title="Combo Point usage breakdown">
-          <ResourceBreakdown
-            tracker={this.comboPointTracker}
-            resourceName="Combo Points"
-            showSpenders={true}
-          />
-        </Tab>
-      ),
-    };
- }
-
-  statisticOrder = STATISTIC_ORDER.CORE(6);
 }
 
-export default ComboPointDetails;
+export default ComboPoints;
